requirements.txt
# Core dependencies for CrewAI + MCP integration
crewai>=0.24.0
crewai-tools[mcp]>=0.6.0
pydantic>=2.4.2

# MCP Protocol support
mcp>=1.0.0

# Environment and utilities
python-dotenv>=1.0.0
loguru>=0.7.0
rich>=13.0.0

# HTTP requests for web tools
requests>=2.31.0
aiohttp>=3.8.0

# Development tools (optional)
pytest>=7.0.0
black>=23.0.0
flake8>=6.0.0

# Additional useful packages
asyncio>=3.4.3
typing-extensions>=4.0.0
